requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
anthropic==0.9.0
pydantic==2.6.0
python-dotenv==1.0.0
httpx==0.26.0
asyncio==3.4.3
